Splits calculateOneBP.pyx and calculateOnebp.pyx into separate files
calculateOneBP contained code to do two things: perform a linear regression
and correct read count biases. This change splits the file into two files,
one for doing the linear regression and one for correcting the read counts.

calculateOnebp contained code to do four things: perform a linear regression,
correct read count biases, calculate covariates for training data and calculate
covariates for non-training data. The code for each of these has been put in its
own file.

This is just a organizational change, no functionality should be affected.
