+++ tests/CorrectBiasStored/getcoefs_test.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 import pyximport; pyximport.install()
 
-from CRADLE.CorrectBiasStored.calculateOneBP import getCoefs
+from CRADLE.CorrectBiasStored.regression import getCoefs
 
 @pytest.mark.parametrize("modelParams,selectedCovariates,result", [
 	([1, 2, 3, 4, 5, 6, 7], [np.nan, 1, np.nan, 1, 1, 1], [1, np.nan, 2, np.nan, 3, 4, 5]),
